lesson 8. Separate classes to files
