+++ src/main/java/lesson_8/MyArrayDataException.java
@@ -1,0 +1,7 @@
+package lesson_8;
+
+public class MyArrayDataException extends Exception {
+    public MyArrayDataException(String message) {
+        super(message);
+    }
+}
