+++ src/main/java/lesson_8/MyArraySizeException.java
@@ -1,0 +1,7 @@
+package lesson_8;
+
+public class MyArraySizeException extends Exception {
+    public MyArraySizeException(String message) {
+        super(message);
+    }
+}
